backend/requirements.txt
Flask==2.0.3
psycopg2-binary
redis
Werkzeug==2.0.3
